+++ source/Public/Edit-CommandDocumentation.ps1
@@ -12,12 +12,9 @@
         Specifies the FileInfo object of the markdown file containing the documentation
         being edited. This parameter is mandatory.
 
-    .PARAMETER SourcePath
-        Specifies the path to the source file of the command. This parameter is mandatory.
-
     .EXAMPLE
-        $filePath = Get-Item -Path "C:\Docs\MyCommand.md"
-        Edit-CommandDocumentation -FilePath $filePath -SourcePath "C:\Scripts\MyCommand.ps1"
+        $filePath = Get-Item -Path 'C:\Docs\MyCommand.md'
+        Edit-CommandDocumentation -FilePath $filePath -SourcePath 'C:\Scripts\MyCommand.ps1'
 
         Edits the documentation for the command specified in the markdown file "C:\Docs\MyCommand.md"
         using the source file "C:\Scripts\MyCommand.ps1".
@@ -30,18 +27,11 @@ function Edit-CommandDocumentation
         [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
         [ValidateScript({Test-Path $_ -PathType 'Leaf'})]
         [System.IO.FileInfo]
-        $FilePath,
-
-        [Parameter(Mandatory = $true)]
-        [ValidateScript({Test-Path $_ -PathType 'Container'})]
-        [System.String]
-        $SourcePath
+        $FilePath
     )
 
     process
     {
-        return
-
         $getMarkdownMetadataScript = @"
 Get-MarkdownMetadata -Path $($FilePath.FullName) -ErrorAction 'Stop'
 "@
@@ -56,7 +46,7 @@ Get-MarkdownMetadata -Path $($FilePath.FullName) -ErrorAction 'Stop'
             The scriptblock is run in a separate process to avoid conflicts with
             other modules that are loaded in the current process.
         #>
-        $markdownMetadata = & $pwshPath -Command $getMarkdownMetadataScriptBlock -ExecutionPolicy 'ByPass'
+        $markdownMetadata = & $pwshPath -Command $getMarkdownMetadataScriptBlock -ExecutionPolicy 'ByPass' -NoProfile
 
         if ($markdownMetadata.Type -ne 'Command')
         {
@@ -65,20 +55,22 @@ Get-MarkdownMetadata -Path $($FilePath.FullName) -ErrorAction 'Stop'
             continue
         }
 
-        $sourceFilePath = $SourcePath | Join-Path -ChildPath "Public/$($FilePath.BaseName).ps1"
-
-        Write-Information -MessageData "Parsing comment based help from source file '$sourceFilePath'." -InformationAction 'Continue'
-
-        $sourceFileCommentBasedHelp = Get-CommentBasedHelp -Path $sourceFilePath -ErrorAction 'Stop'
-
-        write-verbose -message ($sourceFileCommentBasedHelp | Out-String) -Verbose
-
         switch ($markdownMetadata.Schema)
         {
             '2.0.0'
             {
-                # Remove ProgressAction parameter, example code here:
-                # https://github.com/PowerShell/platyPS/issues/595
+                Write-Information -MessageData "Cleaning documentation of '$($FilePath.BaseName)'." -InformationAction 'Continue'
+
+                Remove-MarkdownMetadata -FilePath $FilePath
+
+                <#
+                    Remove ProgressAction parameter from the documentation. The
+                    parameter ProgressAction was introduced in PS 7.4 and is not
+                    supported by PlatyPS <v2.0.0.
+
+                    See issue https://github.com/PowerShell/platyPS/issues/595.
+                #>
+                Remove-ParameterFromMarkdown -FilePath $FilePath -ParameterName 'ProgressAction'
             }
 
             default
@@ -86,7 +78,5 @@ Get-MarkdownMetadata -Path $($FilePath.FullName) -ErrorAction 'Stop'
                 throw "The markdown file '$($FilePath.FullName)' has an unsupported schema version '$($markdownMetadata.Schema)'."
             }
         }
-
-
     }
 }