+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Markdown file.
 - Private function `Remove-ParameterFromMarkdown` that removes a parameter
   from a commands markdown documentation.
+- Public function `Edit-CommandDocumentation` that will modify the a generated
+  command markdown documentation.
 
 ### Changed